admin full done

+++ app/Controllers/Admin.php
@@ -17,18 +17,30 @@ class Admin
         }
     }
 
-    #[NoReturn] public function get($id = null): void
+    #[NoReturn] public static function get($id = null): void
     {
-        Response::send(AdminService::get($id), $id);
+        if (AdminService::isAdmin()) {
+            Response::send(AdminService::get($id), $id);
+        } else {
+            echo('Вы не можете просматривать пользователя так как Вы не авторизовались как администратор');
+        }
     }
 
     #[NoReturn] public function delete($id = null): void
     {
-        Response::send(AdminService::delete($id), $id);
+        if (AdminService::isAdmin()) {
+            Response::send(AdminService::delete($id), $id);
+        } else {
+            echo('Вы не можете удалить пользователя так как Вы не авторизовались как администратор');
+        }
     }
 
-    #[NoReturn] public function update($id, $username, $email, $password, $birthdate, $role): void
+    #[NoReturn] public function update($id = null): void
     {
-        Response::send(AdminService::update($id, $username, $email, $password, $birthdate, $role), $id);
+        if (AdminService::isAdmin()) {
+            Response::send(AdminService::update($id), $id);
+        } else {
+            echo('Вы не можете обновить данные список пользователя так как Вы не авторизовались как администратор');
+        }
     }
 }
